Change the layout of the user part
Now models.js is ApplicationListModel.js
Now configurables.js is splitted into one configurables.js file which contains the models and *.vue files which contains the Vue components of the configurables
