+++ jstests/tests/user/test_application_list_view.js
@@ -1,5 +1,5 @@
 var Vue = require("vuejs");
-var models = require("user/models");
+var ApplicationListModel = require("user/ApplicationListModel");
 var ApplicationListView = require("user/vue-components/ApplicationListView");
 require("filters");
 
@@ -7,7 +7,7 @@ QUnit.module("user.app_list_view");
 QUnit.test("rendering list", function (assert) {
     var done = assert.async();
 
-    var model = new models.ApplicationListModel();
+    var model = new ApplicationListModel();
     var appListView = new ApplicationListView({
         data: function() { return { model: model }; }
     }).$mount();
@@ -39,7 +39,7 @@ QUnit.test("rendering list", function (assert) {
 QUnit.test("rendering nothing in the list", function (assert) {
     var done = assert.async();
 
-    var model = new models.ApplicationListModel();
+    var model = new ApplicationListModel();
     var appListView = new ApplicationListView({
         data: function() { return { model: model }; }
     }).$mount();
@@ -67,7 +67,7 @@ QUnit.test("rendering nothing in the list", function (assert) {
 QUnit.test("search form", function (assert) {
     var done = assert.async();
 
-    var model = new models.ApplicationListModel();
+    var model = new ApplicationListModel();
     var appListView = new ApplicationListView({
         data: function() { return { model: model }; }
     }).$mount();
