+++ jstests/tests/user/test_application_view.js
@@ -1,5 +1,5 @@
 var Vue = require("vuejs");
-var models = require("user/models");
+var ApplicationListModel = require("user/ApplicationListModel");
 var ApplicationView = require("user/vue-components/ApplicationView");
 require("filters");
 
@@ -7,7 +7,7 @@ QUnit.module("user.app_view");
 QUnit.test("rendering form", function (assert) {
     var done = assert.async();
 
-    var model = new models.ApplicationListModel();
+    var model = new ApplicationListModel();
     var appView = new ApplicationView({
         data: function() { return { model: model }; }
     }).$mount();
@@ -40,7 +40,7 @@ QUnit.test("rendering form", function (assert) {
 QUnit.test("rendering iframe", function (assert) {
     var done = assert.async();
 
-    var model = new models.ApplicationListModel();
+    var model = new ApplicationListModel();
     var appView = new ApplicationView({
         data: function() { return { model: model }; }
     }).$mount();
