+++ jstests/tests/user/test_configurables.js
@@ -1,8 +1,9 @@
-var configurables = require("user/configurables");
+var configurables = require("user/configurables/configurables");
+var Resolution = require("user/configurables/Resolution");
 
 QUnit.module("user.configurables");
 QUnit.test("instantiation", function (assert) {
-    var resolutionConf = new configurables.resolution.model();
+    var resolutionConf = new configurables.resolution();
 
     assert.equal(resolutionConf.tag, "resolution");
     assert.deepEqual(resolutionConf.configDict, { resolution: "Window" });
@@ -14,7 +15,7 @@ QUnit.test("instantiation", function (assert) {
 
 QUnit.test("view", function (assert) {
     var propsData = { configDict: { resolution: "Window" } };
-    var component = new configurables.resolution.component({propsData: propsData}).$mount();
+    var component = new Resolution({propsData: propsData}).$mount();
 
     assert.notEqual(component.$el.querySelector("select"), null);
     assert.equal(component.$el.querySelector("select").children.length, 5);
